console: Refactor connect in downlink form
