+++ pkg/webui/locales/en.json
@@ -215,6 +215,16 @@
   "console.components.downlink-form.downlink-form.bytesPayloadDescription": "The desired payload bytes of the downlink message",
   "console.components.downlink-form.downlink-form.jsonPayloadDescription": "The decoded payload of the downlink message",
   "console.components.downlink-form.downlink-form.invalidSessionWarning": "Downlinks can only be scheduled for end devices with a valid session. Please make sure your end device is properly connected to the network.",
+  "console.components.downlink-form.index.insertMode": "Insert Mode",
+  "console.components.downlink-form.index.payloadType": "Payload type",
+  "console.components.downlink-form.index.bytes": "Bytes",
+  "console.components.downlink-form.index.replace": "Replace downlink queue",
+  "console.components.downlink-form.index.push": "Push to downlink queue (append)",
+  "console.components.downlink-form.index.scheduleDownlink": "Schedule downlink",
+  "console.components.downlink-form.index.downlinkSuccess": "Downlink scheduled",
+  "console.components.downlink-form.index.bytesPayloadDescription": "The desired payload bytes of the downlink message",
+  "console.components.downlink-form.index.jsonPayloadDescription": "The decoded payload of the downlink message",
+  "console.components.downlink-form.index.invalidSessionWarning": "Downlinks can only be scheduled for end devices with a valid session. Please make sure your end device is properly connected to the network.",
   "console.components.events.messages.MACPayload": "MAC payload",
   "console.components.events.messages.devAddr": "DevAddr",
   "console.components.events.messages.fPort": "FPort",
@@ -494,6 +504,11 @@
   "console.containers.device-profile-section.device-card.device-card.classA": "Class A",
   "console.containers.device-profile-section.device-card.device-card.classB": "Class B",
   "console.containers.device-profile-section.device-card.device-card.classC": "Class C",
+  "console.containers.device-profile-section.device-card.index.productWebsite": "Product website",
+  "console.containers.device-profile-section.device-card.index.dataSheet": "Data sheet",
+  "console.containers.device-profile-section.device-card.index.classA": "Class A",
+  "console.containers.device-profile-section.device-card.index.classB": "Class B",
+  "console.containers.device-profile-section.device-card.index.classC": "Class C",
   "console.containers.device-profile-section.device-selection.band-select.index.title": "Profile (Region)",
   "console.containers.device-profile-section.device-selection.brand-select.brand-select.title": "End device brand",
   "console.containers.device-profile-section.device-selection.brand-select.brand-select.noOptionsMessage": "No matching brand found",
