+++ pkg/webui/locales/ja.json
@@ -215,6 +215,16 @@
   "console.components.downlink-form.downlink-form.bytesPayloadDescription": "ダウンリンクメッセージの希望ペイロードバイト数",
   "console.components.downlink-form.downlink-form.jsonPayloadDescription": "ダウンリンクメッセージのデコードされたペイロード",
   "console.components.downlink-form.downlink-form.invalidSessionWarning": "ダウンリンクは、有効なセッションを持つエンドデバイスに対してのみスケジュールすることができます。エンドデバイスがネットワークに正しく接続されていることを確認してください",
+  "console.components.downlink-form.index.insertMode": "",
+  "console.components.downlink-form.index.payloadType": "",
+  "console.components.downlink-form.index.bytes": "",
+  "console.components.downlink-form.index.replace": "",
+  "console.components.downlink-form.index.push": "",
+  "console.components.downlink-form.index.scheduleDownlink": "",
+  "console.components.downlink-form.index.downlinkSuccess": "",
+  "console.components.downlink-form.index.bytesPayloadDescription": "",
+  "console.components.downlink-form.index.jsonPayloadDescription": "",
+  "console.components.downlink-form.index.invalidSessionWarning": "",
   "console.components.events.messages.MACPayload": "MAC ペイロード",
   "console.components.events.messages.devAddr": "DevAddr",
   "console.components.events.messages.fPort": "FPort",
@@ -494,6 +504,11 @@
   "console.containers.device-profile-section.device-card.device-card.classA": "",
   "console.containers.device-profile-section.device-card.device-card.classB": "",
   "console.containers.device-profile-section.device-card.device-card.classC": "",
+  "console.containers.device-profile-section.device-card.index.productWebsite": "",
+  "console.containers.device-profile-section.device-card.index.dataSheet": "",
+  "console.containers.device-profile-section.device-card.index.classA": "",
+  "console.containers.device-profile-section.device-card.index.classB": "",
+  "console.containers.device-profile-section.device-card.index.classC": "",
   "console.containers.device-profile-section.device-selection.band-select.index.title": "プロフィール（リージョン）",
   "console.containers.device-profile-section.device-selection.brand-select.brand-select.title": "",
   "console.containers.device-profile-section.device-selection.brand-select.brand-select.noOptionsMessage": "",
